Add degrade fns, move anyi2s and anyi2c
